fix: 캐러셀 코드 정리
