--- client/src/pages/Main/components/SpecialCarousel/SpecialCarousel.module.css
@@ -1,114 +1,0 @@
-.carouselWrapper {
-  position: relative;
-  width: 100%;
-  margin: 2rem 0;
-}
-
-.carouselHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 1rem;
-  padding: 0 1rem;
-}
-
-.carouselTitle {
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: #333;
-}
-
-.moreButton {
-  padding: 0.6rem 1.2rem;
-
-  background-color: transparent;
-
-  font-size: 1.2rem;
-  border: 0.1rem solid #ccc;
-  border-radius: 2rem;
-
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-
-.moreButton:hover {
-  background-color: #f0f0f0;
-}
-
-.carouselViewport {
-  padding: 2rem 0;
-}
-
-/* 공통 슬라이드 */
-.carouselSlide {
-  width: 50% !important;  
-  display: flex;
-  justify-content: center;
-  transition: opacity 0.3s ease, transform 0.3s ease;
-  opacity: 0.3;
-  transform: scale(0.85);
-}
-
-/* 실제 카드 콘텐츠 */
-.slideCard {
-  width: 100%;
-  border-radius: 0.75rem;
-  overflow: hidden;
-  background: white;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-  text-align: center;
-}
-
-/* 이미지 */
-.carouselImage {
-  width: 100%;
-  height: 15rem;
-  object-fit: cover;
-}
-
-/* 텍스트 */
-.cardText {
-  padding: 1rem;
-}
-.cardText h3 {
-  font-size: 1rem;
-  font-weight: bold;
-  margin-bottom: 0.5rem;
-}
-.cardText p {
-  font-size: 0.875rem;
-  color: #666;
-  margin: 0.2rem 0;
-}
-
-/* ✅ 중앙 슬라이드 강조 (실제 크기와 스타일 차이) */
-:global(.swiper-slide-active) {
-  transform: scale(1);
-  opacity: 1;
-}
-
-:global(.swiper-slide-active) .slideCard {
-  transform: scale(1.1);
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2);
-}
-
-/* 버튼 */
-.prevButton,
-.nextButton {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 2.5rem;
-  height: 2.5rem;
-  border-radius: 50%;
-  background: white;
-  border: none;
-  font-size: 1.5rem;
-  z-index: 10;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  cursor: pointer;
-}
-
-.prevButton { left: 1rem; }
-.nextButton { right: 1rem; }
